Move everything which is used for upload and create into a common place
